+++ cypress/support/commands/auth.ts
@@ -38,6 +38,50 @@ function cssRegister() {
     cy.contains('a', 'log in').click();
 }
 
+export function createACLSparql(document: string): string {
+    return `
+        INSERT DATA {
+            @prefix acl: <http://www.w3.org/ns/auth/acl#> .
+
+            <#owner>
+                a acl:Authorization ;
+                acl:agent <${cssPodUrl('/alice/profile/card#me')}>, <mailto:alice@example.com> ;
+                acl:accessTo <${cssPodUrl(`/alice/cookbook/${document}`)}> ;
+                acl:mode acl:Read, acl:Write, acl:Control .
+        }
+    `;
+}
+
+export function publishACLSparql(document: string): string {
+    return `
+        INSERT DATA {
+            @prefix acl: <http://www.w3.org/ns/auth/acl#> .
+            @prefix foaf: <http://xmlns.com/foaf/0.1/> .
+
+            <#public>
+                a acl:Authorization ;
+                acl:agentClass foaf:Agent ;
+                acl:accessTo <${cssPodUrl(`/alice/cookbook/${document}`)}> ;
+                acl:mode acl:Read .
+        }
+    `;
+}
+
+export function unpublishACLSparql(document: string): string {
+    return `
+        DELETE DATA {
+            @prefix acl: <http://www.w3.org/ns/auth/acl#> .
+            @prefix foaf: <http://xmlns.com/foaf/0.1/> .
+
+            <#public>
+                a acl:Authorization ;
+                acl:agentClass foaf:Agent ;
+                acl:accessTo <${cssPodUrl(`/alice/cookbook/${document}`)}> ;
+                acl:mode acl:Read .
+        }
+    `;
+}
+
 export function cssPodUrl(path: string = ''): string {
     return `http://localhost:4000${path}`;
 }